Extracted common tests
