+++ test/integration_common_test.go
@@ -13,6 +13,7 @@ import (
 	"encoding/json"
 	"encoding/pem"
 	"fmt"
+	"github.com/gruntwork-io/terragrunt/cli/commands/terraform"
 	"io"
 	"math/big"
 	"net"
@@ -315,4 +316,51 @@ func certSetup(t *testing.T) (*tls.Config, *tls.Config) {
 	}
 
 	return serverTLSConf, clientTLSConf
+}
+
+func validateOutput(t *testing.T, outputs map[string]TerraformOutput, key string, value interface{}) {
+	t.Helper()
+	output, hasPlatform := outputs[key]
+	assert.Truef(t, hasPlatform, "Expected output %s to be defined", key)
+	assert.Equalf(t, output.Value, value, "Expected output %s to be %t", key, value)
+}
+
+// wrappedBinary - return which binary will be wrapped by Terragrunt, useful in CICD to run same tests against tofu and terraform
+func wrappedBinary() string {
+	value, found := os.LookupEnv("TERRAGRUNT_TFPATH")
+	if !found {
+		// if env variable is not defined, try to check through executing command
+		if util.IsCommandExecutable(tofuBinary, "-version") {
+			return tofuBinary
+		}
+		return terraformBinary
+	}
+	return filepath.Base(value)
+}
+
+// expectedWrongCommandErr - return expected error message for wrong command
+func expectedWrongCommandErr(command string) error {
+	if wrappedBinary() == tofuBinary {
+		return terraform.WrongTofuCommand(command)
+	}
+	return terraform.WrongTerraformCommand(command)
+}
+
+func isTerraform() bool {
+	return wrappedBinary() == terraformBinary
+}
+
+func findFilesWithExtension(dir string, ext string) ([]string, error) {
+	var files []string
+	err := filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+		if !info.IsDir() && filepath.Ext(path) == ext {
+			files = append(files, path)
+		}
+		return nil
+	})
+
+	return files, err
 }
